backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pandas==2.1.4
sqlalchemy==2.0.23
google-generativeai==0.3.0
python-dotenv==1.0.0
pydantic==2.5.3
